move formatNumber function into and export from StatsBar
